CR-1127851 Refactor memory tag function into xocl_xclbin
