+++ src/runtime_src/core/pcie/driver/linux/xocl/xocl_types.h
@@ -1,0 +1,21 @@
+/*
+ * Copyright (C) 2022 Advanced Micro Devices, Inc. All rights reserved.
+ *
+ * Authors: daniel.benusovich@amd.com
+ *
+ * This software is licensed under the terms of the GNU General Public
+ * License version 2, as published by the Free Software Foundation, and
+ * may be copied, distributed, and modified under those terms.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#ifndef	_XOCL_TYPES_H_
+#define	_XOCL_TYPES_H_
+
+typedef	void *xdev_handle_t;
+
+#endif
